Pixel Fehler gefixt

+++ Game2/Items/Item.cs
@@ -115,7 +115,7 @@ namespace Game1
                     sourceRectangle: sourceRect,
                     color: new Color(r * illuminationStrength, g * illuminationStrength, b * illuminationStrength, alpha),
                     rotation: angle,
-                    origin: new Vector2(sourceRect.Width * 0.5f, sourceRect.Height * 0.5f),
+                     origin: new Vector2((float)sourceRect.Width * 0.5f, (float)sourceRect.Height * 0.5f),
                     scale: new Vector2(1 * (size.X / (float)sourceRect.Width), 1 * (size.Y / (float)sourceRect.Height)),
                     effects: SpriteEffects.None,
                     layerDepth: 1);
